Rename sync_blog to contentful
Also remove all references to the blog stuff
