+++ lib/contentful_middleman.rb
@@ -2,6 +2,6 @@ require "middleman-core"
 
 require 'contentful_middleman/version'
 require 'contentful_middleman/core'
-require "contentful_middleman/commands/sync_blog"
+require "contentful_middleman/commands/contentful"
 
 ::Middleman::Extensions.register(:contentful, ContentfulMiddleman::Core)
